feat: get data from ODRÉ and insert it in influxdb

+++ influxdb.py
@@ -1,11 +1,10 @@
 from datetime import timezone, datetime
-import influxdb_client
-from influxdb_client import WriteApi, QueryApi, DeleteApi, Point
+from influxdb_client import InfluxDBClient, WriteApi, QueryApi, DeleteApi, Point
 from influxdb_client.client.write_api import SYNCHRONOUS
 
 g_bucket = ''
 g_org = ''
-
+g_client: InfluxDBClient | None = None
 g_write_api: WriteApi | None = None
 g_read_api: QueryApi | None = None
 g_delete_api: DeleteApi | None = None
@@ -14,24 +13,28 @@ g_delete_api: DeleteApi | None = None
 def init_influxdb(bucket, org, token, url):
     global g_bucket
     global g_org
+    global g_client
     global g_write_api
     global g_read_api
     global g_delete_api
 
-    client = influxdb_client.InfluxDBClient(
+    g_client = InfluxDBClient(
         url=url,
         token=token,
         org=org)
 
     g_bucket = bucket
     g_org = org
-    g_write_api = client.write_api(write_options=SYNCHRONOUS)
-    g_read_api = client.query_api()
-    g_delete_api = client.delete_api()
+    g_write_api = g_client.write_api(write_options=SYNCHRONOUS)
+    g_read_api = g_client.query_api()
+    g_delete_api = g_client.delete_api()
 
 
-def is_not_connected():
-    return (g_write_api is None) and (g_read_api is None) and (g_delete_api is None)
+def is_connected():
+    return g_client.ping() and \
+        (g_write_api is not None) and \
+        (g_read_api is not None) and \
+        (g_delete_api is not None)
 
 
 class PointData:
@@ -41,12 +44,16 @@ class PointData:
         self.value = value
 
 
-def write(measurement_name: str, fields: list[PointData]):
-    if is_not_connected():
+def write(measurement_name: str, timestamp: datetime | None, fields: list[PointData]):
+    if not is_connected():
         print('ERROR: InfluxDB not initialized')
         return
 
     p = Point(measurement_name)
+
+    if timestamp is not None:
+        p = p.time(timestamp.isoformat())
+
     for field in fields:
         if field.is_field:
             p = p.field(field.key, field.value)
@@ -57,12 +64,12 @@ def write(measurement_name: str, fields: list[PointData]):
 
 
 def read_all(measurement_name: str):
-    if is_not_connected():
+    if not is_connected():
         print('ERROR: InfluxDB not initialized')
         return
 
     query = f'from(bucket: "{g_bucket}")' \
-            f'|> range(start: 0)' \
+            f'|> range(start: 1970-01-01T00:00:00Z)' \
             f'|> filter(fn: (r) => r._measurement == "{measurement_name}")' \
             f'|> last()'
 
@@ -77,11 +84,13 @@ def read_all(measurement_name: str):
 
 
 def delete_all(measurement_name: str):
-    date_now = datetime.now(tz=timezone.utc).isoformat()
+    if not is_connected():
+        print('ERROR: InfluxDB not initialized')
+        return
 
     g_delete_api.delete(
         start='1970-01-01T00:00:00Z',
-        stop=date_now,
+        stop=datetime.now(tz=timezone.utc).isoformat(),
         predicate=f'_measurement="{measurement_name}"',
         bucket=g_bucket,
         org=g_org)
